#4 Style: 통합 테스트 이름 수정
